! consolidate some standard search functions
